+++ karma.conf.js
@@ -31,7 +31,6 @@ module.exports = function(config) {
 
     // list of files / patterns to load in the browser
     files: [
-      require.resolve('regenerator-runtime/runtime'),
       {
         pattern: 'tests/test-karma.js',
         watched: false, served: true, included: true
@@ -68,11 +67,12 @@ module.exports = function(config) {
               options: {
                 presets: ['@babel/preset-env'],
                 plugins: [
-                  '@babel/plugin-transform-regenerator',
                   [
                     '@babel/plugin-proposal-object-rest-spread',
                     {useBuiltIns: true}
-                  ]
+                  ],
+                  '@babel/plugin-transform-modules-commonjs',
+                  '@babel/plugin-transform-runtime'
                 ]
               }
             }